Add POC flow action for gathering acq package info
Add action to gather acq. package info by using a script step, 
dot-walking method, and JSON streaming API. The dot-walking
method appears to be the simplest for gathering data when 
there is a direct reference from one record to a different record
in another table. However, if there is not a reference from the 
base record (e.g., acq. package --> portfolio) then this method
will not work and an additional query will be required.

Ticket: AT-7315