Add POC database view for gathering data
Add a database view with the acquisition package and
provisioning job tables. A  where clause is used to match
the provisioning job with the acquisition package using
an id.

This successfully provides records with access to both 
the acquisition packages and the associated provisioning 
jobs, which was confirmed through a simple background 
script using GlideRecord to get records in the Database
View created.

Ticket: AT-7315