Update action to use database view record
Use GlideRecord in the action to perform a query on 
the database view to find the acq package that is 
provided to the action. The action then provides data
from both the acq package and provision job records 
as outputs (this assumes a provisioning job is 
associated with the acq package).

Ticket: AT-7315